+++ package.json
@@ -64,8 +64,7 @@
     "react-native": ">=0.70.0"
   },
   "dependencies": {
-    "@expo/config": "^8.1.2 || ^9.0.1",
-    "@expo/config-plugins": "^7.2.5 || ^8.0.4",
+    "@expo/config-plugins": "^8.0.4",
     "detect-indent": "^6.1.0",
     "fs-extra": "^11.2.0",
     "node-html-parser": "^6.1.13",
@@ -81,6 +80,7 @@
     "@types/fs-extra": "^11.0.4",
     "@types/node": "^20.12.7",
     "@types/react": "^18.2.60",
+    "@types/semver": "^7.5.8",
     "@types/sharp": "^0.31.1",
     "@typescript-eslint/eslint-plugin": "^7.8.0",
     "@typescript-eslint/parser": "^7.8.0",
@@ -90,6 +90,7 @@
     "react": "18.2.0",
     "react-native": "0.74.3",
     "react-native-builder-bob": "^0.23.2",
+    "semver": "^7.6.2",
     "typescript": "^5.4.5"
   },
   "codegenConfig": {
